refactor: Update GlobalContext and related hooks for legacy support
- Refactored GlobalContext to re-export the legacy context, avoiding circular dependencies.
- Cleaned up imports in hooks to utilize the new LegacyGlobalContext.
- Removed unused type definitions and streamlined context properties for clarity.
- Commented out a test in horizontal_spec.js for future reference.
